relocate use cases test add get coordinate by ip use case
