+++ src/main/kotlin/logic/usecases/clothessuggester/SuggestClothesUseCase.kt
@@ -1,4 +1,4 @@
-package logic.usecases.clothesSuggester
+package logic.usecases.clothessuggester
 
 import logic.model.Cloth
 import logic.model.ClothType
