+++ src/test/kotlin/ui/clothessuggester/ClothesSuggesterByCityNameUITest.kt
@@ -7,7 +7,7 @@ import io.mockk.verify
 import kotlinx.coroutines.ExperimentalCoroutinesApi
 import logic.model.Cloth
 import logic.model.ClothType
-import logic.usecases.clothesSuggester.GetClothingSuggestionUseCase
+import logic.usecases.clothessuggester.GetClothingSuggestionUseCase
 import logic.usecases.location.GetCoordinateByCityAndCountryUseCase
 import logic.usecases.weather.GetDailyWeatherByCoordinateUseCase
 import ui.ClothesSuggesterByCityNameUI
